+++ src/main/java/stl/threebodysimulation/Particle.java
@@ -6,34 +6,34 @@ import java.util.HashMap;
 
 public class Particle {
     // Attributes of a particle. Arrays represent vectors, with index 0 being x-coordinate and 1 being y-coordinate.
-    int[] position;
-    int[] velocity;
-    int[] acceleration;
-    final int mass;
-    final int id;
+    double[] position;
+    double[] velocity;
+    double[] acceleration;
+    final double mass;
+    final byte id;
     final Color color;
 
     // Attributes packaged for sending to the InfoFXMLController.
-    HashMap<String, int[]> packagedInformation;
+    HashMap<String, double[]> packagedInformation;
 
-    Particle(int xPos, int yPos, int xVel, int yVel, int mass, int id, Color color) {
+    Particle(double xPos, double yPos, double xVel, double yVel, double mass, byte id, Color color) {
         // Constructor for the Particle class.
         // INPUTS:
-        // xPos: int, the x-component of the particle's position.
-        // yPos: int, the y-component of the particle's position.
-        // xVel: int, the x-component of the particle's velocity.
-        // yVel: int, the y-component of the particle's velocity.
-        // mass: int, the mass of the particle.
-        // id: int, the id of the particle, one of {1, 2, 3}
+        // xPos: double, the x-component of the particle's position.
+        // yPos: double, the y-component of the particle's position.
+        // xVel: double, the x-component of the particle's velocity.
+        // yVel: double, the y-component of the particle's velocity.
+        // mass: double, the mass of the particle.
+        // id: byte, the id of the particle, one of {1, 2, 3}
         // color: Color, the color of the particle as a javafx Color object.
 
         // USAGE:
         // >> Particle(0, 0, -5, 3, 10, 1, Color.BLUE)
         // creates a blue particle that represents object 1 with position [0, 0], velocity [-5, 3], and mass 10.
 
-        position = new int[]{xPos, yPos};
-        velocity = new int[]{xVel, yVel};
-        acceleration = new int[]{0, 0}; // users cannot provide starting acceleration to a particle.
+        position = new double[]{xPos, yPos};
+        velocity = new double[]{xVel, yVel};
+        acceleration = new double[]{0, 0}; // users cannot provide starting acceleration to a particle.
 
         // Packages above vectors into convenient hashmap for later use.
         packagedInformation = new HashMap<>();
